Cleared Deriver Method & Math Fixes
This commit fixes all of the issues I accidentally created when trying to implement FE recognition.

- If the Deriver block is grabbed and the energy is extracted by a value at or above the max production rate it will clear the fortron inside of the device.

- Added documentation to the Fortron helper functions

+++ src/main/java/dev/su5ed/mffs/blockentity/FortronBlockEntity.java
@@ -63,7 +63,7 @@ public abstract class FortronBlockEntity extends InventoryBlockEntity implements
     /**
      * @return The initial fortron tank capacity in buckets
      */
-    public int getBaseFortronTankCapacity() {
+    public double getBaseFortronTankCapacity() {
         return 1;
     }
 
@@ -71,7 +71,7 @@ public abstract class FortronBlockEntity extends InventoryBlockEntity implements
      * @return The base fortron tank capacity multiplied by the forge bucket volume
      */
     public int getRealFortronTankCapacity() {
-        return getBaseFortronTankCapacity() * FluidType.BUCKET_VOLUME;
+        return (int) (getBaseFortronTankCapacity() * FluidType.BUCKET_VOLUME);
     }
 
     protected List<ItemStack> getCards() {
